First implementation of write support. It seames that is does not work...
Ported all relevant functions from inode.c and balloc.c without
prealloc / reservation support. For now the code just builds and
hopefully not breaks anything else. Overwriting seams to work though

+++ src/jext2/BlockAccess.java
@@ -23,8 +23,8 @@ public class BlockAccess {
 		BlockAccess.instance = this;
 	}
 
-	/** Read a block offsize specified by setBlocksize() at logical address nr */
-	public ByteBuffer read(int nr) throws IOException {
+	/** Read a block off size specified by setBlocksize() at logical address nr */
+	public ByteBuffer read(long nr) throws IOException {
 		ByteBuffer buf = ByteBuffer.allocate(blocksize);		
 		buf.order(ByteOrder.BIG_ENDIAN);
